SkupperKube CLI implementation
