+++ cmd/skupper/skupper_token.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"context"
-	"fmt"
 	"time"
 
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
@@ -10,7 +8,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/skupperproject/skupper/api/types"
-	"github.com/skupperproject/skupper/pkg/utils"
 )
 
 func NewCmdToken() *cobra.Command {
@@ -26,7 +23,7 @@ var password string
 var expiry time.Duration
 var uses int
 
-func NewCmdTokenCreate(newClient cobraFunc, flag string) *cobra.Command {
+func NewCmdTokenCreate(skupperClient SkupperClient, flag string) *cobra.Command {
 	subflag := ""
 	if flag == "client-identity" {
 		subflag = "i"
@@ -39,33 +36,8 @@ func NewCmdTokenCreate(newClient cobraFunc, flag string) *cobra.Command {
 		Use:    "create <output-token-file>",
 		Short:  "Create a token.  The 'link create' command uses the token to establish a link from a remote Skupper site.",
 		Args:   cobra.ExactArgs(1),
-		PreRun: newClient,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			silenceCobra(cmd)
-			switch tokenType {
-			case "cert":
-				err := cli.ConnectorTokenCreateFile(context.Background(), clientIdentity, args[0])
-				if err != nil {
-					return fmt.Errorf("Failed to create token: %w", err)
-				}
-				return nil
-			case "claim":
-				name := clientIdentity
-				if name == "skupper" {
-					name = ""
-				}
-				if password == "" {
-					password = utils.RandomId(24)
-				}
-				err := cli.TokenClaimCreateFile(context.Background(), name, []byte(password), expiry, uses, args[0])
-				if err != nil {
-					return fmt.Errorf("Failed to create token: %w", err)
-				}
-				return nil
-			default:
-				return fmt.Errorf("invalid token type. Specify cert or claim")
-			}
-		},
+		PreRun: skupperClient.NewClient,
+		RunE:   skupperClient.TokenCreate,
 	}
 	cmd.Flags().StringVarP(&clientIdentity, flag, subflag, types.DefaultVanName, "Provide a specific identity as which connecting skupper installation will be authenticated")
 	cmd.Flags().StringVarP(&tokenType, "token-type", "t", "claim", "Type of token to create. Valid options are 'claim' or 'cert'")
